+++ R/corpus_sample.R
@@ -59,6 +59,7 @@ corpus_sample.corpus <- function(x, size = NULL, replace = FALSE, prob = NULL, b
 
     if (!missing(by)) {
         by <- eval(substitute(by), get_docvars(x, user = TRUE, system = TRUE), parent.frame())
+        if (is.factor(by)) by <- droplevels(by)
     }
 
     i <- resample(seq_len(ndoc(x)), size = size, replace = replace, prob = prob, by = by)