+++ tests/testthat/test-corpus.R
@@ -68,7 +68,7 @@ test_that("corpus constructors works for kwic", {
       kwic("econom*", window = 10, separator = "") %>%
       corpus(split_context = FALSE, extract_keyword = FALSE)
     expect_identical(
-        texts(corp)[2],
+        as.character(corp)[2],
         c("text1.L390" = "it is decimating the domestic economy? As we are tired ")
     )
 
@@ -77,7 +77,7 @@ test_that("corpus constructors works for kwic", {
     toks1 <- tokens(txt1, what = "word", remove_separators = FALSE)
     kw1 <- kwic(toks1, "a", window = 10, separator = "")
     expect_equivalent(
-        texts(corpus(kw1, split_context = FALSE)), txt1
+        as.character(corpus(kw1, split_context = FALSE)), txt1
     )
 
     # quotes
@@ -85,25 +85,25 @@ test_that("corpus constructors works for kwic", {
     toks2 <- tokens(txt2, what = "word", remove_separators = FALSE)
     kw2 <- kwic(toks2, "a", window = 10, separator = "")
     expect_equivalent(
-        texts(corpus(kw2, split_context = FALSE)), txt2
+        as.character(corpus(kw2, split_context = FALSE)), txt2
     )
     txt3 <- "This \"is\" only a test!"
     toks3 <- tokens(txt3, what = "word", remove_separators = FALSE)
     kw3 <- kwic(toks3, "a", window = 10, separator = "")
     expect_equivalent(
-        texts(corpus(kw3, split_context = FALSE)), txt3
+        as.character(corpus(kw3, split_context = FALSE)), txt3
     )
     txt4 <- 'This "is" only (a) test!'
     toks4 <- tokens(txt4, what = "word", remove_separators = FALSE)
     kw4 <- kwic(toks4, "a", window = 10, separator = "")
     expect_equivalent(
-        texts(corpus(kw4, split_context = FALSE)), txt4
+        as.character(corpus(kw4, split_context = FALSE)), txt4
     )
     txt5 <- "This is only (a) test!"
     toks5 <- tokens(txt5, what = "word", remove_separators = FALSE)
     kw5 <- kwic(toks5, "a", window = 10, separator = "")
     expect_equivalent(
-        texts(corpus(kw5, split_context = FALSE)), txt5
+        as.character(corpus(kw5, split_context = FALSE)), txt5
     )
     expect_error(corpus(kw, split_context = logical(), extract_keyword = FALSE))
     expect_error(corpus(kw, extract_keyword = logical()))
@@ -158,7 +158,7 @@ test_that("test corpus constructor works for tm objects", {
     # VCorpus
     data(crude, package = "tm")    # load in a tm example VCorpus
     corp1 <- corpus(crude)
-    expect_equal(substring(texts(corp1)[1], 1, 21),
+    expect_equal(substring(as.character(corp1)[1], 1, 21),
                  c("reut-00001.xml"  = "Diamond Shamrock Corp"))
 
     data(acq, package = "tm")
@@ -170,7 +170,7 @@ test_that("test corpus constructor works for tm objects", {
     scorp <- SimpleCorpus(DirSource(txt, encoding = "UTF-8"),
                                 control = list(language = "lat"))
     corp3 <- corpus(scorp)
-    expect_equal(content(scorp), texts(corp3))
+    expect_equal(content(scorp), as.character(corp3))
 
     # any other type
     scorp2 <- scorp
@@ -188,7 +188,7 @@ test_that("test corpus constructor works for VCorpus with one document (#445)", 
     require(tm)
     vcorp <- VCorpus(VectorSource(data_corpus_inaugural[2]))
     corp <- corpus(vcorp)
-    expect_equivalent(texts(corp)[1], texts(data_corpus_inaugural)[2])
+    expect_equivalent(as.character(corp)[1], as.character(data_corpus_inaugural)[2])
     detach("package:tm", unload = FALSE, force = TRUE)
     detach("package:NLP", unload = FALSE, force = TRUE)
 })
@@ -252,7 +252,7 @@ test_that("corpus constructor works with tibbles", {
         "corpus"
     )
     expect_equal(
-        texts(corpus(dd)),
+        as.character(corpus(dd)),
         c(text1 = "Hello", text2 = "quanteda", text3 = "world")
     )
 })
@@ -378,7 +378,7 @@ test_that("corpus.data.frame sets docnames correctly", {
 test_that("corpus handles NA correctly (#1372, #1969)", {
     txt <- c("a b c", NA, "d e f")
     expect_true(!any(
-        suppressWarnings(is.na(texts(corpus(txt)))
+        suppressWarnings(is.na(as.character(corpus(txt)))
     )))
     expect_warning(
         corpus(txt),
@@ -386,7 +386,7 @@ test_that("corpus handles NA correctly (#1372, #1969)", {
     )
     expect_true(!any(
         suppressWarnings(
-          is.na(texts(corpus(data.frame(text = txt, stringsAsFactors = FALSE))))
+          is.na(as.character(corpus(data.frame(text = txt, stringsAsFactors = FALSE))))
         )
     ))
 })
@@ -440,7 +440,7 @@ test_that("handle data.frame with improper column names and text and doc_id fiel
 
     expect_equal(names(docvars(corp)), c("V1", "dvar2", "V3"))
     expect_equal(docnames(corp), paste0("txt", 1:5))
-    expect_equivalent(texts(corp), LETTERS[1:5])
+    expect_equivalent(as.character(corp), LETTERS[1:5])
 })
 
 test_that("handle data.frame variable renaming when one already exists", {