+++ tests/testthat/test-dfm.R
@@ -9,25 +9,6 @@ test_that("test c.corpus", {
 })
 
 ## rbind.dfm
-## TODO: Test classes returned
-
-test_that("test rbind.dfm with the same columns", {
-    fox <- "What does the fox say?"
-    foxdfm <- rep(1, 20)
-    dim(foxdfm) <- c(4, 5)
-    colnames(foxdfm) <- c("does", "fox", "say", "the", "what")
-    rownames(foxdfm) <-  rep(c("text1", "text2"), 2)
-
-    suppressWarnings(dfm1 <- dfm(tokens(c(fox, fox)), remove_punct = TRUE))
-
-    expect_true(
-        all(rbind(dfm1, dfm1) == foxdfm)
-    )
-    expect_that(
-        rbind(dfm1, dfm1),
-        is_a("dfm")
-    )
-})
 
 # TODO: Add function for testing the equality of dfms
 
@@ -476,24 +457,6 @@ test_that("dfm's document counts in verbose message is correct", {
                    "kept 2 features")
 })
 
-test_that("dfm head, tail work as expected", {
-    dfmt <- head(data_dfm_lbgexample, 4, nf = 3)
-    expect_equal(featnames(dfmt), LETTERS[1:3])
-    expect_equal(docnames(dfmt), paste0("R", 1:4))
-
-    dfmt <- head(data_dfm_lbgexample, -4, nf = -30)
-    expect_equal(featnames(dfmt), LETTERS[1:7])
-    expect_equal(docnames(dfmt), paste0("R", 1:2))
-
-    dfmt <- tail(data_dfm_lbgexample, 4, nf = 3)
-    expect_equal(featnames(dfmt), c("ZI", "ZJ", "ZK"))
-    expect_equal(docnames(dfmt), c("R3", "R4", "R5", "V1"))
-
-    dfmt <- tail(data_dfm_lbgexample, -4, nf = -34)
-    expect_equal(featnames(dfmt), c("ZI", "ZJ", "ZK"))
-    expect_equal(docnames(dfmt), c("R5", "V1"))
-})
-
 test_that("dfm print works with options as expected", {
     dfmt <- dfm(tokens(data_corpus_inaugural[1:14],
                 remove_punct = FALSE, remove_numbers = FALSE, split_hyphens = TRUE))
@@ -596,7 +559,7 @@ test_that("dfm works with stem options", {
         c("run", "ran")
     )
     expect_error(
-        dfm(tokens(txt_english), stem = c(TRUE, FALSE)),
+        suppressWarnings(dfm(tokens(txt_english), stem = c(TRUE, FALSE))),
         "The length of stem must be 1"
     )